[refactor] diary 앱 urls, views, serializers 폴더로 리팩토링
