+++ diary/serialziers/image_serializers.py
@@ -1,0 +1,5 @@
+from diary.validator import *
+
+
+class ImageUrlRequest(serializers.Serializer):
+    imgUrl = serializers.CharField()
