+++ diary/urls/quiz_urls.py
@@ -1,5 +1,6 @@
 from django.urls import path
-from .views import *
+
+from diary.views.quiz_views import GetQuizView
 
 urlpatterns = [
     path('', GetQuizView.as_view()),
