+++ diary/utils/text_rank/stop_words.py
