+++ image/views.py
@@ -10,7 +10,8 @@ from rest_framework.views import APIView
 
 from config.basemodel import ApiResponse, validator
 from config.settings import REQUEST_QUERY, JWT_SECRET, REQUEST_BODY
-from diary.serializers import *
+from diary.serialziers.diary_serializers import *
+from diary.serialziers.keyword_serializers import FindKeywordImgRequest
 from image.gpt.GenerateImage import generate_upload_image, test_generate_image_urls
 from image.s3_modules.s3_handler import upload_file_to_s3
 from image.serializers import *
